+++ packages/uring_api/src/_uring_api/uring_api_send_all.c
@@ -10,7 +10,6 @@
 #include "uring_api_park.h"
 #include "uring_api_prepare.h"
 #include "uring_api_probe.h"
-#include "uring_api_staging.h"
 
 #ifndef IORING_RECVSEND_POLL_FIRST
 #define IORING_RECVSEND_POLL_FIRST (1U << 0)